refactor(server): modularize PR feedback workflow runtime (GH-1649)
Mechanically extract the 1,677-line workflow_runtime_pr_feedback.rs
into three narrowly owned child modules, keeping the root file as the
stable facade for all nine external caller files:

- persistence.rs: private persist_* functions, merge approval
  persistence, commit_runtime_decision, RuntimeDecisionCommitOutcome
- command_state.rs: active-command queries, command-status predicates,
  failed-child suppression
- targets.rs: definition registration, issue/PR target loading,
  instance construction, workflow IDs, runtime JSON and field helpers

Every function and type moved as a complete body with pub(super)
visibility only where sibling or root access requires it; moved private
helpers used by the nested test suite stay reachable through the root
namespace via explicit #[cfg(test)] use imports, so no test file
changes. Pre/post inventory: 47 functions and 10 types, exact multiset
match, every symbol owned exactly once. All touched production files
are within the 800-line ceiling (root 511, persistence 755, targets
309, command_state 170).

Closes #1649
